requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
alembic==1.13.2
    # via -r requirements.in
amqp==5.2.0
    # via kombu
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
argon2-cffi==23.1.0
    # via pwdlib
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
async-timeout==4.0.3
    # via asyncpg
asyncpg==0.29.0
    # via -r requirements.in
bcrypt==4.1.2
    # via pwdlib
billiard==4.2.0
    # via celery
black==24.8.0
    # via -r requirements.in
celery==5.4.0
    # via
    #   -r requirements.in
    #   flower
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
cffi==1.17.0
    # via
    #   argon2-cffi-bindings
    #   cryptography
click==8.1.7
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   typer
    #   uvicorn
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via
    #   click
    #   pytest
    #   uvicorn
cryptography==43.0.0
    # via pyjwt
dnspython==2.6.1
    # via email-validator
email-validator==2.1.2
    # via
    #   fastapi
    #   fastapi-users
fastapi[all]==0.112.0
    # via
    #   -r requirements.in
    #   fastapi-users
fastapi-cli[standard]==0.0.5
    # via fastapi
fastapi-users[sqlalchemy]==13.0.0
    # via
    #   -r requirements.in
    #   fastapi-users-db-sqlalchemy
fastapi-users-db-sqlalchemy==6.0.1
    # via fastapi-users
flower==2.0.1
    # via -r requirements.in
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via fastapi
humanize==4.10.0
    # via flower
idna==3.7
    # via
    #   anyio
    #   email-validator
    #   httpx
iniconfig==2.0.0
    # via pytest
itsdangerous==2.2.0
    # via fastapi
jinja2==3.1.4
    # via fastapi
kombu==5.4.0
    # via celery
makefun==1.15.4
    # via fastapi-users
mako==1.3.5
    # via alembic
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   jinja2
    #   mako
mdurl==0.1.2
    # via markdown-it-py
mypy-extensions==1.0.0
    # via black
orjson==3.10.7
    # via fastapi
packaging==24.1
    # via
    #   black
    #   pytest
pathspec==0.12.1
    # via black
platformdirs==4.2.2
    # via black
pluggy==1.5.0
    # via pytest
prometheus-client==0.20.0
    # via flower
prompt-toolkit==3.0.47
    # via click-repl
psycopg2==2.9.9
    # via -r requirements.in
pwdlib[argon2,bcrypt]==0.2.0
    # via fastapi-users
pycparser==2.22
    # via cffi
pydantic==2.8.2
    # via
    #   fastapi
    #   pydantic-extra-types
    #   pydantic-settings
pydantic-core==2.20.1
    # via pydantic
pydantic-extra-types==2.9.0
    # via fastapi
pydantic-settings==2.4.0
    # via fastapi
pygments==2.18.0
    # via rich
pyjwt[crypto]==2.8.0
    # via fastapi-users
pytest==8.3.2
    # via
    #   -r requirements.in
    #   pytest-asyncio
pytest-asyncio==0.24.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via celery
python-dotenv==1.0.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.9
    # via
    #   fastapi
    #   fastapi-users
pytz==2024.1
    # via flower
pyyaml==6.0.2
    # via
    #   fastapi
    #   uvicorn
redis==5.0.8
    # via -r requirements.in
rich==13.7.1
    # via typer
shellingham==1.5.4
    # via typer
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlalchemy[asyncio]==2.0.32
    # via
    #   -r requirements.in
    #   alembic
    #   fastapi-users-db-sqlalchemy
starlette==0.37.2
    # via fastapi
tornado==6.4.1
    # via flower
typer==0.12.3
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer
tzdata==2024.1
    # via celery
ujson==5.10.0
    # via fastapi
uvicorn[standard]==0.30.5
    # via
    #   fastapi
    #   fastapi-cli
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
watchfiles==0.23.0
    # via uvicorn
wcwidth==0.2.13
    # via prompt-toolkit
websockets==12.0
    # via uvicorn
